requirements.txt
numpy
pandas
PyYAML
nibabel
tqdm
scipy
scikit-image
torch
